Bump max number of connections for the server
4 is quite low, doesn't support very many different consumers and
consumer pools. 60 is the default from the NIOServerCnxnFactory source
code so we'll use that instead.

It might be handy to pull this variable out into the "kafka-config" map
as well if it needs to be adjustable from the outside world. Leaving it
as a hardcoded constant at the moment as it might not be necessary.